eval: move EvalFleet to miles/ray/rollout/eval_fleet.py
checkpoint_eval.py's own docstring stopped matching its contents when the fleet
left the CheckpointEvalFn contract: it still said FleetEvalFn was the contract's
in-job implementation. The file held two unrelated things - a 36-line user-facing
contract and 99 lines of engine plumbing - and someone writing a backend had to
read past _pin_fleet, _mark_unreachable_engines and _wait_router_ready to reach
the part that concerns them.

miles/ray/rollout/ is where it belongs by what it touches: RolloutServer.recover,
wait_all_engines_alive, engine actor handles, the eval router's address. It sat
in miles/rollout/ only because it was born as a CheckpointEvalFn. The file is
named after the class, as rollout_server.py / server_group.py / eval_dispatch.py
are.

checkpoint_eval.py is now 98 lines and reads as one thing: the contract, EvalSkip,
retarget_args (public for external backends), and the posture predicates.
eval_fleet.py imports EvalSkip and retarget_args from it; nothing goes back the
other way. The three fleet tests move to tests/fast/ray/rollout/test_eval_fleet.py
alongside the module, using that directory's conftest make_args.
